Add files via upload
I intentionally decided to include a key performance indicator (KPI) in the dataset via creating a measure using DAX in powerBI. i was just testing how to use the performance indicator.
I made use of filters and slicers to enable an easier way of visualizing the datasets. The dataset consists of countries, product segments, discount band and so on. 
with the use of cards, slicers, filters pie charts and others as well the drill down buttons and other functions used, we can know which country, region, product is doing well and also those struggling and then further research and analysis from collected datasets in these regions can make the company improve on areas the datasets points to.